RSMS-BE-DOCKERIZED/backend/RSMS/requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
inflection==0.5.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.7.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
